final changes

+++ CartModule/src/main/java/com/cts/service/ICartService.java
@@ -10,7 +10,7 @@ public interface ICartService {
 	CartDTO addProductToCart(Integer userId, ProductDTO productdto,Long bookId);
 	CartDTO increaseProductQuantity(Integer userId, Long bookId, Integer quantityToAdd);
 	CartDTO decreaseProductQuantity(Integer userId, Long bookId, Integer quantityToRemove);
-    void removeProductFromCart(Integer userId, Integer bookId);
+    void removeProductFromCart(Integer userId, Long bookId);
     void clearCart(Integer userId);
     double calculateTotalPrice(Integer userId);
     CartDTO createCartForUser(Integer userId);
